Attempt to fix #2
Now only check the candidate circle against the placed circles other than the 2 used to determine the candidate intead of includin them. It was leading to numeric errors that would prevent good candidates to be taken into account in the solution.

Now throws an exception when a value cannot be placed instead of loging.

Also:
- added HedgeTestCase class
- revert import of sqrt and pi in the global namespace
- use machine eps to generate candidate wich leads to slightly better solution density.
- Other minor refactoring.