Added ex={"datum": <input value>} dictionary when input is scalar
Scalar input data like [0, 1, 2 ...] would lose the information related to the input values when converted to Circle even after we added the ex dictionary. The input values is now stored in the ex dict even when the input is a simple scalar value. That allows for downstream to display sensible text defaults when drawing the circles.

Also:
- Enhanced the bubbles helper function to make labels optional and default labels to the id, then the datum and finally a simple count of the circle.
- Updated test to reflect changes to ex field.
- Added pyreadline to the dev-requirements for convenience.